requirements.txt
fastapi
uvicorn
langchain-google-genai
langgraph
python-dotenv